Add bootstrap and remove app-network from compose file.
In order to setup the project for development, this commit adds a
`script/bootstrap` file to install dependencies and add an `.env` file
to use with docker setup.  This commit also removes the `app-network`
from the `docker-compose.yml` file as `docker-compose` creates a default
network that is bridged by default.  To verify run `docker network ls`
after you have run `script/bootstrap`. This commit fixes issue #4.